Added first piece of PCA for RGB. Just need to add methods to store mapped colors and report it to MSHeaderViewController so that the method call for PCA creation can be completed

+++ Spectral_Motion/MSBandMappingTableVC.m
@@ -164,7 +164,7 @@
         m_BandsSelected = [[NSMutableArray alloc]init];
         for(NSIndexPath *path in [self.tableView indexPathsForSelectedRows])
         {
-            [m_BandsSelected setObject:[NSNumber numberWithInt:path.row] atIndexedSubscript:count] ;
+            [m_BandsSelected setObject:[NSNumber numberWithInt:(int)path.row] atIndexedSubscript:count] ;
             count++;
         }
